fastapi/requirements.txt
aioredis==2.0.1
fastapi==0.89.1
orjson==3.8.5
pydantic==1.10.4
uvicorn==0.20.0
uvloop==0.17.0
python-dotenv==0.21.1
aiokafka==0.8.0
clickhouse-driver==0.2.5
mypy==0.991
sentry-sdk[fastapi]==1.15.0
motor==3.1.1